chore: col and row are moved to the layout folder
