--- src/col/col.en-US.md
@@ -1,16 +1,0 @@
-:: BASE_DOC ::
-
-## API
-
-### Col Props
-
-name | type | default | description | required
--- | -- | -- | -- | --
-offset | String / Number | - | \- | N
-span | String / Number | - | \- | N
-
-### Row Props
-
-name | type | default | description | required
--- | -- | -- | -- | --
-gutter | String / Number | - | Spacing between columns (default unit px) | N
