--- src/col/col.md
@@ -1,15 +1,0 @@
-:: BASE_DOC ::
-
-## API
-### Col Props
-
-名称 | 类型 | 默认值 | 说明 | 必传
--- | -- | -- | -- | --
-offset | String / Number | - | 列的偏移量（默认单位px） | N
-span | String / Number | - | 列的宽度（默认单位px） | N
-
-### Row Props
-
-名称 | 类型 | 默认值 | 说明 | 必传
--- | -- | -- | -- | --
-gutter | String / Number | - | 列之间的间距（默认单位px） | N
