--- src/col/style/index.js
@@ -1,1 +1,0 @@
-import '../../_common/style/mobile/components/col/v2/_index.less';
