+++ src/layout/col-props.ts
@@ -3,18 +3,17 @@
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
+
+import { TdColProps } from '../layout/type';
 import { PropType } from 'vue';
-import { TdColProps } from './type';
-const props = {
-  /** 列之间的间距（默认单位px） */
+
+export default {
+  /** 列的偏移量（默认单位px） */
   offset: {
     type: [String, Number] as PropType<TdColProps['offset']>,
   },
-
-  /** 列之间的间距（默认单位px） */
+  /** 列的宽度（默认单位px） */
   span: {
     type: [String, Number] as PropType<TdColProps['span']>,
   },
 };
-
-export default props;
