+++ src/layout/layout.en-US.md
@@ -1,0 +1,18 @@
+:: BASE_DOC ::
+
+## API
+
+
+### Row Props
+
+name | type | default | description | required
+-- | -- | -- | -- | --
+gutter | String / Number | - | \- | N
+
+
+### Col Props
+
+name | type | default | description | required
+-- | -- | -- | -- | --
+offset | String / Number | - | \- | N
+span | String / Number | - | \- | N
