+++ src/layout/layout.md
@@ -1,0 +1,18 @@
+:: BASE_DOC ::
+
+## API
+
+
+### Row Props
+
+名称 | 类型 | 默认值 | 描述 | 必传
+-- | -- | -- | -- | --
+gutter | String / Number | - | 列之间的间距（默认单位px） | N
+
+
+### Col Props
+
+名称 | 类型 | 默认值 | 描述 | 必传
+-- | -- | -- | -- | --
+offset | String / Number | - | 列的偏移量（默认单位px） | N
+span | String / Number | - | 列的宽度（默认单位px） | N
