+++ src/layout/row.tsx
@@ -1,6 +1,6 @@
 import { computed, defineComponent, CSSProperties, provide } from 'vue';
 import { convertUnit } from '../shared';
-import props from './props';
+import props from './row-props';
 import config from '../config';
 import { rowInjectionKey } from './constants';
 import { useTNodeJSX } from '../hooks/tnode';
