+++ src/layout/style/index.js
@@ -1,0 +1,1 @@
+import '../../_common/style/mobile/components/layout/_index.less';
