+++ src/layout/type.ts
@@ -4,13 +4,20 @@
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
 
-export interface TdColProps {
+export interface TdRowProps {
   /**
    * 列之间的间距（默认单位px）
    */
-  offset: string | number;
+  gutter?: string | number;
+}
+
+export interface TdColProps {
+  /**
+   * 列的偏移量（默认单位px）
+   */
+  offset?: string | number;
   /**
    * 列的宽度（默认单位px）
    */
-  span: string | number;
+  span?: string | number;
 }
