--- src/row/row.en-US.md
@@ -1,9 +1,0 @@
-:: BASE_DOC ::
-
-## API
-
-### Row Props
-
-name | type | default | description | required
--- | -- | -- | -- | --
-gutter | String / Number | - | \- | N
