--- src/row/row.md
@@ -1,11 +1,0 @@
-:: BASE_DOC ::
-
-## 引入
-
-## API
-
-### Row Props
-
-名称 | 类型 | 默认值 | 说明 | 必传
--- | -- | -- | -- | --
-gutter | String / Number | - | 列之间的间距（默认单位px） | N
