--- src/row/style/index.js
@@ -1,1 +1,0 @@
-import '../../_common/style/mobile/components/row/v2/_index.less';
